requirements.txt
torch>=1.13
torchvision
numpy
scikit-learn
openslide-python
Pillow
tqdm
flwr>=1.6
opacus>=1.1.0
scikit-image
faiss-cpu
opencv-python
matplotlib
pandas

